chore: demoを整備
